Add dynamic mode switch/actions and help link
Add javascript detecting the value of the first character of the query
string.

When '@' is found, the mode will be set to 'command'
When '#' is found, the mode will be set to 'bookmark'
When '!' is found, the mode will be set to 'bang'
When '\' is found, the mode will be set to 'ddg1st'

If one of these modes is activated a [cancel] link will be displayed to
allow setting it back to default.

When string is empty and none of the above prefix characters are used,
the mode will be set from 'unset' to 'default' after typing the first
character.